% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NarxNetwork.R
\name{CoFESNARX}
\alias{CoFESNARX}
\title{CoFESNARX}
\usage{
CoFESNARX(x, y, dim, KL2 = 0.01, BL2 = 0.01, dr_o = 0, act = "relu")
}
\arguments{
\item{y, x}{numeric vectors for the target and exegenous variables}

\item{dim}{numberic vector for the dimension of the deisred netowrk}

\item{KL2}{the L2 kernel_regularizerregularization factor for hidden layer kernal regularization}

\item{BL2}{the L2 bias_regularizer regularization factor for hidden layer bias regularization}

\item{dr_o}{drop out rate used between hidden layers}

\item{act}{activation funciton used, if not specified 'relu' used}
}
\value{
returns the target combined with the original and lagged exegenous
 variables in one database.
}
\description{
\code{CoFESNARX} returns a compiled NARX model via Keras
}
\details{
This function takes target variable series, exogenous variables series, and the number of desired
 lags for each group respectively and then returns the requested lag series for each in a single
 data set.  Ydelay/Xdelay are vectors that holds the consecutive delays required for the target
 variable/exogenous series (ex 1:3 or 4:5)
}
