% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/narxdataprep.R
\name{CoFESNARXdata}
\alias{CoFESNARXdata}
\title{CoFESNARXdata}
\usage{
CoFESNARXdata(yy, xx, ydelay, xdelay, delay = 1, date = T)
}
\arguments{
\item{yy}{numeric vectors containing target data}

\item{xx}{numeric vectors containing feature data}

\item{ydelay}{integer representing the desired AR order for y}

\item{xdelay}{integer representing the desired AR order for x}

\item{delay}{integers representing n, the size of the prediction window into the future}

\item{date}{boolean value to indicate if you data includes dates}
}
\value{
returns the target combined with the original and lagged exegenous variables in one database.
}
\description{
\code{CoFESNARXdata} returns the lead/lag series necessary to run the CoFESNARX
}
\details{
This function transforms the target and exegenous series variables for the
 CoFESNARX model.  variables series, and the number of desired lags for each
 group respectively and then returns the requested lag series for each in a
 single  data set.  Ydelay/Xdelay are vectors that holds the consecutive
 delays required for the target variable series (ex 1:3 or 4:5).
}
