% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/narxdataprep.R
\name{returnTARGET}
\alias{returnTARGET}
\title{returnTARGET}
\usage{
returnTARGET(data)
}
\arguments{
\item{data}{numeric vectors}
}
\value{
returns the target
}
\description{
\code{returnTARGET} returns TARGET created by CoFESNARXdata/CoFESSNARXdata
}
\details{
This function returns the target series created by CoFESNARXdata and CoFESSNARXdata.
}
