% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/narxdataprep.R
\name{returnX}
\alias{returnX}
\title{returnX}
\usage{
returnX(data, date = T)
}
\arguments{
\item{data}{numeric vectors}

\item{date}{Boolen if date is included}
}
\value{
returns the exogenous series
}
\description{
\code{returnX} returns exogenous series
}
\details{
This function returns the exogenous series created by CoFESNARXdata and CoFESSNARXdata.
}
